tegra-state-scripts: remove the correct file
When updating the nv_boot_control.conf file during the payload update the wrong file was removed. The "${mnt}/etc/nv_boot_control.conf" file is a symbolic link to "${mnt}/var/lib/nvbootctrl/nv_boot_control.conf". When the symbolic link in /etc is removed, the next firmware update will fail.

Instead of removing the symbolic link, the file it points to is removed and recreated after the setup-nv-boot-control script is executed.

Signed-off-by: Boris Nagels <develop@focusware.nl>