meta-mender-tegra: update flash layout file for jetson-nano-qspi-sd
The flashing tools and flash layout for the QSPI-equipped Nanos were
updated in L4T R32.4.3, and our existing layout file generates an
error during USB flashing because the updated tools think the VER
partition runs off the end of the flash. Either we added up the
sizes incorrectly, or there's a bug in the tools.

This patch updates the flash layout to keep the U-Boot environment
area we reserve in the QSPI flash at the same location and with the
same size.  The VER_b and VER partitions that follow are reduced
in size to prevent the flashing tools from complaining.  This should
retain compatibility for an OTA upgrade that updates the SPI flash
without having to save and restore the U-Boot environment.  The
VER partitions are smaller than the default L4T settings (which
were also reduced in R32.4.3), but still plenty large enough to
hold the version information.

Signed-off-by: Matt Madison <matt@madison.systems>