mender-custom-flash-layout: remove partition numbers
With L4T R32.4.4, the `id` attribute on a partition
is used to specify the partition number.  This allows,
for instance, the APP partition to be physically located
at the end of an SDcard but appear as partition 1 in the
GPT.

Since our custom layouts were derived from an earlier
version of L4T that simply used `id` as a sequence number
for the flashing tools with no external effect, we must
remove those attributes to be compatible with the newer
tools.

Signed-off-by: Matt Madison <matt@madison.systems>